requirements.txt
# ── Async runtime ───────────────────────────────────────────────────────────
asyncio-throttle>=1.0.2

# ── Polymarket CLOB SDK ─────────────────────────────────────────────────────
py-clob-client>=0.17.0

# ── Web3 / EIP-712 signing ──────────────────────────────────────────────────
web3>=6.15.0
eth-account>=0.11.0
eth-typing>=4.0.0

# ── WebSocket ───────────────────────────────────────────────────────────────
websockets>=12.0

# ── HTTP client ─────────────────────────────────────────────────────────────
aiohttp>=3.9.0

# ── Database drivers ────────────────────────────────────────────────────────
aiosqlite>=0.20.0
asyncpg>=0.29.0          # optional – only needed for PostgreSQL backend

# ── NTP ─────────────────────────────────────────────────────────────────────
ntplib>=0.4.0

# ── CLI dashboard ───────────────────────────────────────────────────────────
rich>=13.7.0

# ── Environment helpers ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ───────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
